move search to form fields
