+++ projects/ui-framework/src/lib/form-elements/index.ts
@@ -1,5 +1,4 @@
 export * from './datepicker';
 export * from './input';
 export * from './textarea';
-export * from './select-elements/single-select';
-export * from './select-elements/multi-select';
+export * from './select-elements';
