+++ projects/ui-framework/src/lib/form-elements/search/index.ts
@@ -1,0 +1,1 @@
+export * from './search.module';
