+++ projects/ui-framework/src/lib/form-elements/search/search.component.scss
