+++ projects/ui-framework/src/lib/form-elements/select-elements/index.ts
@@ -1,0 +1,3 @@
+export * from './select/select.interface';
+export * from './single-select/single-select.module';
+export * from './multi-select/multi-select.module';
