--- projects/ui-framework/src/lib/form-elements/select-elements/multi-select/index.ts
@@ -1,2 +1,0 @@
-export * from '../select/select.interface';
-export * from './multi-select.module';
