+++ projects/ui-framework/src/lib/form-elements/select-elements/multi-select/multi-select.component.ts
@@ -7,8 +7,8 @@ import { SelectGroupOption } from '../select/select.interface';
 })
 export class MultiSelectComponent {
 
-  @Input() options: SelectGroupOption[];
-  @Input() selectedIds: (string | number)[];
+  @Input() options: SelectGroupOption[] = [];
+  @Input() selectedIds: (string | number)[] = [];
   @Input() showSingleGroupHeader = false;
 
   constructor() {
