+++ projects/ui-framework/src/lib/form-elements/select-elements/select/select.component.spec.ts
@@ -5,7 +5,7 @@ import { CommonModule } from '@angular/common';
 import { NoopAnimationsModule } from '@angular/platform-browser/animations';
 import { SelectComponent } from './select.component';
 import { SelectModelService } from './select-model.service';
-import { SearchModule } from '../../../search/search.module';
+import { SearchModule } from '../../search/search.module';
 import { ButtonsModule } from '../../../buttons';
 import { OverlayContainer } from '@angular/cdk/overlay';
 import { Platform } from '@angular/cdk/platform';
