+++ projects/ui-framework/src/lib/form-elements/select-elements/select/select.module.ts
@@ -3,7 +3,7 @@ import { SelectComponent } from './select.component';
 import { MatFormFieldModule, MatInputModule, MatOptionModule, MatSelectModule } from '@angular/material';
 import { CommonModule } from '@angular/common';
 import { FormsModule } from '@angular/forms';
-import { SearchModule } from '../../../search/search.module';
+import { SearchModule } from '../../search/search.module';
 import { ButtonsModule } from '../../../buttons';
 import { SelectModelService } from './select-model.service';
