--- projects/ui-framework/src/lib/form-elements/select-elements/single-select/index.ts
@@ -1,2 +1,0 @@
-export * from '../select/select.interface';
-export * from './single-select.module';
